+++ CHANGELOG.md
@@ -7,6 +7,15 @@ Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [Unreleased]
+
+### Changed
+
+- Reorganised source files into a Python package (`fsic`):
+    - `fsic_fortran` is now `fsic.fortran`
+    - `fsictools` is now `fsic.tools`
+
+
 ## [0.7.1.dev] - 2021-08-04
 
 Various additions and changes to improve consistency and checks (both parser